requirements.txt
flask
numpy
opencv-python
keras
tensorflow
pillow
spotipy
